+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## Unreleased
+## [0.4.2] - 2023-03-11
 
 ### Added
 
@@ -135,6 +135,7 @@ The two most important changes in usage are:
 [0.3.6]: https://github.com/dogmatiq/ferrite/releases/tag/v0.3.6
 [0.4.0]: https://github.com/dogmatiq/ferrite/releases/tag/v0.4.0
 [0.4.1]: https://github.com/dogmatiq/ferrite/releases/tag/v0.4.1
+[0.4.2]: https://github.com/dogmatiq/ferrite/releases/tag/v0.4.2
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD